Working on add_constant_t_apex_pressure_chromo_set_depth.pro. This code will use a set number of cells based on n_depth.  The step size will exponentially increase from the corona until the depth is reached.
2012-08-01 Changes in the HyLoop Tree

Changed shrec_conduction to allow ensure that temp1 and temp2 are never negative.
Changed add_chromo_heat so that when the SET_SYS keyword is set the background coronal heating and the chromospheric heating is smoothed to avoid too sharp of a derivative.

Removed uneeded unit_tests/conduction/sphere_axes.gif image.